requirements.txt
youtube-transcript-api>=0.6.0
openai>=0.27.0
python-dotenv>=1.0.0
